requirements.txt
Django==4.1.6
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
